+++ src/utils.ts
@@ -4,8 +4,6 @@ import { exec } from 'child_process';
 import readline from 'readline';
 import { AxePuppeteer } from '@axe-core/puppeteer';
 import puppeteer from 'puppeteer';
-import { runCodeScan } from './CodeBaseScan';
-import { CreateAssistant } from './ModelHandler';
 
 interface ImproveSemanticsOptions {
   htmlFilePath: string;
@@ -14,89 +12,6 @@ interface ImproveSemanticsOptions {
   openVSCode?: boolean;
 }
 
-// Start Generation Here
-async function init() {
-  console.log('Starting init function');
-
-  const apiKey = await GetUserAPIKey();
-
-  try {
-    const contextFiles = await runCodeScan();
-
-    const assistant = await CreateAssistant(apiKey, contextFiles);
-
-    console.log('Assistant "GuideDog" created successfully:');
-
-    await UpdateConfig(assistant);
-
-    return assistant;
-  } catch (error) {
-    console.error('Error creating assistant:', error);
-    throw error;
-  } finally {
-    console.log('Init function completed');
-  }
-}
-
-async function Check() {
-  // Axe-Core
-}
-
-async function FixFile() {}
-
-async function FixRepo() {}
-
-async function UpdateConfig(assistant: OpenAI.Beta.Assistants.Assistant) {
-  // Read existing config or create a new one
-  let config: { assistantId: string } = { assistantId: '' }; // TODO: make this a proper config object
-
-  try {
-    const existingConfig = await fs.readFile('guidedog.config.js', 'utf8');
-    config = JSON.parse(existingConfig);
-  } catch (error) {
-    console.log('No existing config found, creating a new one.');
-  }
-
-  // Append assistantId to the config
-  config['assistantId'] = assistant.id;
-
-  // Write the updated config back to the file
-  await fs.writeFile(
-    'guidedog.config.js',
-    `module.exports = ${JSON.stringify(config, null, 2)};`,
-    'utf8',
-  );
-  console.log('Configuration saved to guidedog.config.js');
-}
-
-async function GetUserAPIKey(): Promise<string> {
-  const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout,
-  });
-
-  let apiKey = '';
-  const originalStderr = process.stderr.write.bind(process.stderr);
-
-  try {
-    // Temporarily suppress stderr
-    process.stderr.write = () => true; // Suppress stderr output
-
-    apiKey = await new Promise<string>((resolve) => {
-      rl.question('Enter your OpenAI API key: ', (answer) => {
-        resolve(answer.trim());
-        rl.close();
-      });
-    });
-  } catch (error) {
-    console.error('Error getting API key:', error);
-    rl.close();
-  } finally {
-    // Restore stderr
-    process.stderr.write = originalStderr;
-    return apiKey;
-  }
-}
 //Old Functions
 async function improveHtmlSemantics({
   htmlFilePath,
@@ -242,4 +157,4 @@ const assessAccessibility = async (flag: boolean) => {
 };
 
 // Export functions for use as a module
-export { improveSemantics, improveHtmlSemantics, init, assessAccessibility };
+// export { improveSemantics, improveHtmlSemantics, init, assessAccessibility };
