+++ src/utils/check.ts
@@ -1,0 +1,3 @@
+export async function check(flag: boolean){
+  console.log(`check${flag? ' --report': ''}`)
+}
