+++ src/utils/fix.ts
@@ -1,0 +1,7 @@
+export async function fixFile() {
+  console.log('fix specific file')
+}
+
+export async function fixRepo() {
+  console.log('fix the repo')
+}
